+++ src/piratechain/PiratechainTools.ts
@@ -12,7 +12,7 @@ import {
   EdgeWalletInfo,
   JsonObject
 } from 'edge-core-js/types'
-import { Tools as ToolsType } from 'react-native-piratechain'
+import { base16, base64 } from 'rfc4648'
 
 import { PluginEnvironment } from '../common/innerPlugin'
 import { asIntegerString } from '../common/types'
@@ -32,7 +32,7 @@ export class PiratechainTools implements EdgeCurrencyTools {
   currencyInfo: EdgeCurrencyInfo
   io: EdgeIo
   networkInfo: PiratechainNetworkInfo
-  nativeTools: typeof ToolsType
+  piratechainIo: PiratechainIo
 
   constructor(env: PluginEnvironment<PiratechainNetworkInfo>) {
     const { builtinTokens, currencyInfo, io, networkInfo } = env
@@ -48,7 +48,7 @@ export class PiratechainTools implements EdgeCurrencyTools {
       throw new Error('Need piratechain native IO')
     }
 
-    this.nativeTools = piratechainIo.Tools
+    this.piratechainIo = piratechainIo
   }
 
   async getDisplayPrivateKey(
@@ -65,14 +65,11 @@ export class PiratechainTools implements EdgeCurrencyTools {
   }
 
   async getNewWalletBirthdayBlockheight(): Promise<number> {
-    return await this.nativeTools.getBirthdayHeight(
-      this.networkInfo.rpcNode.defaultHost,
-      this.networkInfo.rpcNode.defaultPort
-    )
+    return await this.piratechainIo.getLatestNetworkHeight()
   }
 
   async isValidAddress(address: string): Promise<boolean> {
-    return await this.nativeTools.isValidAddress(address)
+    return await this.piratechainIo.isValidAddress(address)
   }
 
   // will actually use MNEMONIC version of private key
@@ -143,13 +140,17 @@ export class PiratechainTools implements EdgeCurrencyTools {
     if (typeof mnemonic !== 'string') {
       throw new Error('InvalidMnemonic')
     }
-    const unifiedViewingKey: string = await this.nativeTools.deriveViewingKey(
+
+    // Registers the wallet with the SDK's registry as a side effect,
+    // using the same alias name the engine looks up later:
+    const viewingKey = await this.piratechainIo.deriveViewingKey({
+      name: base16.stringify(base64.parse(walletInfo.id)),
       mnemonic,
-      this.networkInfo.rpcNode.networkName
-    )
+      birthdayHeight: piratechainPrivateKeys.birthdayHeight
+    })
     return {
       birthdayHeight: piratechainPrivateKeys.birthdayHeight,
-      publicKey: unifiedViewingKey
+      publicKey: viewingKey
     }
   }